refactor(extension): split grammar.rs god-file into modules
